# parallax_v1_3_artifact_pack/contracts/openapi/parallax_api_v1_3.yaml
openapi: 3.1.0
info:
  title: Parallax API
  version: 1.3.0
  description: Canonical v1.3 API contract for Parallax, a temporal-first personal
    intelligence application with timing analytics and user-controlled context awareness.
servers:
- url: https://api.parallax.local
  description: Local/staging base URL
security:
- bearerAuth: []
paths:
...
  /v1/activities:
    post:
      summary: Create activity
      operationId: createActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      summary: List activities
      operationId: listActivities
      parameters:
      - name: q
        in: query
        schema:
          type: string
      - name: limit
        in: query
        schema:
          type: integer
          default: 50
          maximum: 100
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Activity'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/activities/resolve:
    post:
      summary: Resolve activity name or alias
      description: Read-only resolver. This endpoint must not create aliases, emit
        events, enqueue jobs, or write audit rows beyond normal request logs. Alias
        creation uses /v1/activities/{activity_id}/aliases and requires a mutation
        envelope.
      operationId: resolveActivity
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResolveActivityRequest'
      responses:
        '200':
          description: Resolution candidates
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResolveActivityResponse'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/activities/{activity_id}:
    get:
      summary: Get activity
      operationId: getActivity
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Activity'
        '404':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/activities/{activity_id}/aliases:
    get:
      summary: List activity aliases
      operationId: listActivityAliases
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Activity aliases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityAlias'
    post:
      summary: Add activity alias
      operationId: addActivityAlias
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddActivityAliasRequest'
      responses:
        '201':
          description: Alias created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityAlias'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/activities/{activity_id}/aliases/{alias_id}/decision:
    post:
      summary: Decide activity alias suggestion
      operationId: decideActivityAlias
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: alias_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecideActivityAliasRequest'
      responses:
        '200':
          description: Alias after decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityAlias'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/activities/{activity_id}/relationships:
    get:
      summary: List activity relationships
      operationId: listActivityRelationships
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Activity relationships
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ActivityRelationship'
    post:
      summary: Create activity relationship
      operationId: createActivityRelationship
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateActivityRelationshipRequest'
      responses:
        '201':
          description: Relationship created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityRelationship'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/activities/{activity_id}/relationships/{relationship_id}/decision:
    post:
      summary: Decide activity relationship suggestion
      operationId: decideActivityRelationship
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: relationship_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecideActivityRelationshipRequest'
      responses:
        '200':
          description: Relationship after decision
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityRelationship'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/activities/{activity_id}/merge-preview:
    post:
      summary: Preview activity merge
      description: Read-only preview. This endpoint must not mutate activity identity,
        create audit rows, emit events, or enqueue jobs.
      operationId: previewActivityMerge
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityMergePreviewRequest'
      responses:
        '200':
          description: Merge preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityMergePreview'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/activities/{activity_id}/merge:
    post:
      summary: Merge activity into another activity
      operationId: mergeActivities
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivityMergeRequest'
      responses:
        '200':
          description: Activity identity change
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityIdentityChange'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/activities/{activity_id}/split-preview:
    post:
      summary: Preview activity split
      description: Read-only preview. This endpoint defines the Phase 6 split UX
        contract without committing a split.
      operationId: previewActivitySplit
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ActivitySplitPreviewRequest'
      responses:
        '200':
          description: Split preview
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivitySplitPreview'
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /v1/activities/{activity_id}/profile:
    get:
      summary: Get activity profile
      operationId: getActivityProfile
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityProfile'
        '404':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
  /v1/activities/{activity_id}/checkpoints:
    get:
      summary: List checkpoint templates
      operationId: listCheckpoints
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Checkpoint templates
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
    put:
      summary: Replace checkpoint templates
      operationId: putCheckpoints
      parameters:
      - name: activity_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PutCheckpointsRequest'
      responses:
        '200':
          description: Checkpoint templates
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
        '400':
          description: Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
...
components:
  schemas:
    MutationEnvelope:
      type: object
      additionalProperties: false
      properties:
        client_mutation_id:
          type: string
          minLength: 1
        client_device_id:
          type: string
          minLength: 1
        client_timestamp:
          type: string
          format: date-time
        idempotency_key:
          type: string
          minLength: 1
        client_sequence:
          type:
          - integer
          - 'null'
          minimum: 0
      required:
      - client_mutation_id
      - client_device_id
      - client_timestamp
      - idempotency_key
    ApiError:
      type: object
      additionalProperties: false
      properties:
        error_code:
          type: string
        message:
          type: string
        details:
          type: object
        request_id:
          type: string
        retryable:
          type: boolean
        docs_ref:
          type:
          - string
          - 'null'
      required:
      - error_code
      - message
      - request_id
      - retryable
...
    Activity:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        display_name:
          type: string
          minLength: 1
        canonical_key:
          type:
          - string
          - 'null'
        description:
          type:
          - string
          - 'null'
        status:
          type: string
          enum:
          - active
          - archived
          - merged
        merged_into_activity_id:
          type:
          - string
          - 'null'
          format: uuid
        default_timing_mode:
          type: string
          enum:
          - estimate_only
          - whole_task
          - checkpointed
          - routine
          - calibration
          - passive
        privacy_class:
          type: string
          enum:
          - normal
          - sensitive
          - private
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required:
      - id
      - user_id
      - display_name
      - status
      - default_timing_mode
      - privacy_class
      - created_at
      - updated_at
...
    ActivityAlias:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        activity_id:
          type: string
          format: uuid
        alias_text:
          type: string
        normalized_alias:
          type: string
        source:
          type: string
          enum:
          - user
          - system_suggested
          - imported
        confidence:
          type: number
          minimum: 0
          maximum: 1
        user_confirmed:
          type: boolean
        rejected:
          type: boolean
        created_at:
          type: string
          format: date-time
      required:
      - id
      - user_id
      - activity_id
      - alias_text
      - normalized_alias
      - source
      - confidence
      - user_confirmed
      - rejected
      - created_at
    ActivityRelationship:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        from_activity_id:
          type: string
          format: uuid
        to_activity_id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
          - same_as
          - alias_of
          - part_of
          - has_checkpoint
          - variant_of
          - related_to
          - usually_before
          - usually_after
        metadata:
          type: object
        user_confirmed:
          type: boolean
        state:
          type: string
          enum:
          - suggested
          - confirmed
          - rejected
        created_at:
          type: string
          format: date-time
      required:
      - id
      - user_id
      - from_activity_id
      - to_activity_id
      - kind
      - metadata
      - user_confirmed
      - state
      - created_at
    ActivityMergePreview:
      type: object
      additionalProperties: false
      properties:
        source_activity_id:
          type: string
          format: uuid
        target_activity_id:
          type: string
          format: uuid
        relationship_kind:
          type: string
          enum:
          - same_as
        affected_session_count:
          type: integer
          minimum: 0
        history_preservation:
          type: string
          enum:
          - source_activity_soft_merged
        requires_confirmation:
          type: boolean
      required:
      - source_activity_id
      - target_activity_id
      - relationship_kind
      - affected_session_count
      - history_preservation
      - requires_confirmation
    ActivityIdentityChange:
      type: object
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        change_type:
          type: string
          enum:
          - merge
        source_activity_id:
          type: string
          format: uuid
        target_activity_id:
          type: string
          format: uuid
        affected_session_count:
          type: integer
          minimum: 0
        audit_id:
          type: string
          format: uuid
        created_at:
          type: string
          format: date-time
      required:
      - id
      - user_id
      - change_type
      - source_activity_id
      - target_activity_id
      - affected_session_count
      - audit_id
      - created_at
    ActivitySplitPreview:
      type: object
      additionalProperties: false
      properties:
        source_activity_id:
          type: string
          format: uuid
        proposed_display_name:
          type: string
        movable_session_count:
          type: integer
          minimum: 0
        requires_confirmation:
          type: boolean
        commit_supported:
          type: boolean
      required:
      - source_activity_id
      - proposed_display_name
      - movable_session_count
      - requires_confirmation
      - commit_supported
...
    CreateActivityRequest:
      type: object
      additionalProperties: false
      properties:
        mutation:
          $ref: '#/components/schemas/MutationEnvelope'
        display_name:
          type: string
          minLength: 1
        description:
          type:
          - string
          - 'null'
        default_timing_mode:
          type: string
          enum:
          - estimate_only
          - whole_task
          - checkpointed
          - routine
          - calibration
          - passive
          default: whole_task
        privacy_class:
          type: string
          enum:
          - normal
          - sensitive
          - private
          default: normal
      required:
      - mutation
      - display_name
    ResolveActivityRequest:
      type: object
      additionalProperties: false
      properties:
        query:
          type: string
          minLength: 1
        limit:
          type: integer
          minimum: 1
          maximum: 25
          default: 10
      required:
      - query
    ResolveActivityResponse:
      type: object
      additionalProperties: false
      properties:
        candidates:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              activity:
                oneOf:
                - $ref: '#/components/schemas/Activity'
                - type: 'null'
              display_name:
                type: string
              confidence:
                type: number
                minimum: 0
                maximum: 1
              match_type:
                type: string
                enum:
                - canonical
                - alias
                - fuzzy
                - no_match
              evidence:
                type: object
            required:
            - display_name
            - confidence
            - match_type
            - evidence
        recommended_activity_id:
          type:
          - string
          - 'null'
          format: uuid
        requires_confirmation:
          type: boolean
      required:
      - candidates
      - requires_confirmation
...
    AddActivityAliasRequest:
      type: object
      additionalProperties: false
      properties:
        mutation:
          $ref: '#/components/schemas/MutationEnvelope'
        alias_text:
          type: string
          minLength: 1
        user_confirmed:
          type: boolean
          default: true
      required:
      - mutation
      - alias_text
    DecideActivityAliasRequest:
      type: object
      additionalProperties: false
      properties:
        mutation:
          $ref: '#/components/schemas/MutationEnvelope'
        decision:
          type: string
          enum:
          - accept
          - reject
        reason:
          type:
          - string
          - 'null'
      required:
      - mutation
      - decision
    CreateActivityRelationshipRequest:
      type: object
      additionalProperties: false
      properties:
        mutation:
          $ref: '#/components/schemas/MutationEnvelope'
        to_activity_id:
          type: string
          format: uuid
        kind:
          type: string
          enum:
          - same_as
          - alias_of
          - part_of
          - has_checkpoint
          - variant_of
          - related_to
          - usually_before
          - usually_after
        metadata:
          type: object
        user_confirmed:
          type: boolean
          default: true
      required:
      - mutation
      - to_activity_id
      - kind
    DecideActivityRelationshipRequest:
      type: object
      additionalProperties: false
      properties:
        mutation:
          $ref: '#/components/schemas/MutationEnvelope'
        decision:
          type: string
          enum:
          - accept
          - reject
        reason:
          type:
          - string
          - 'null'
      required:
      - mutation
      - decision
    ActivityMergePreviewRequest:
      type: object
      additionalProperties: false
      properties:
        target_activity_id:
          type: string
          format: uuid
      required:
      - target_activity_id
    ActivityMergeRequest:
      type: object
      additionalProperties: false
      properties:
        mutation:
          $ref: '#/components/schemas/MutationEnvelope'
        target_activity_id:
          type: string
          format: uuid
        reason:
          type:
          - string
          - 'null'
      required:
      - mutation
      - target_activity_id
    ActivitySplitPreviewRequest:
      type: object
      additionalProperties: false
      properties:
        proposed_display_name:
          type: string
          minLength: 1
        session_ids:
          type: array
          items:
            type: string
            format: uuid
      required:
      - proposed_display_name
    PutCheckpointsRequest:
      type: object
      additionalProperties: false
      properties:
        mutation:
          $ref: '#/components/schemas/MutationEnvelope'
        checkpoints:
          type: array
          items:
            type: object
            additionalProperties: false
            properties:
              label:
                type: string
              sequence_order:
                type: integer
                minimum: 1
              optional:
                type: boolean
                default: false
              phase_type:
                type:
                - string
                - 'null'
            required:
            - label
            - sequence_order
      required:
      - mutation
      - checkpoints
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT